// src/types/region-analysis.d.ts
export interface RegionStatistics {
  region_name: string;
  population: number;
  total_visit_count: number;
  total_cost: number;
  chojin_rejin_visit_count: number;
  sinhwan_visit_count: number;
  average_cost_per_visit: number;
  average_cost_per_patient: number;
  total_patient_count: number;
  average_patient_age: number;
}